Timehandling: choose safer representation for fractional seconds (closes #939)
When drafting the time handling framework some years ago,
I foresaw the possible danger of mixing up numbers relating
to fractional seconds, with other plain numbers intended as
frame counts or as micro ticks. Thus I deliberately picked
an incompatible integer type for FSecs = boost::rational<long>

However, using long is problematic in itself, since its actual
bit length is not fixed, and especially on 32bit platforms long
is quite surprisingly defined to be the same as int.

However, meanwhile, using the new C++ features, I have blocked
pretty much any possible implicit conversion path, requiring
explicit conversions in the relevant ctor invocations. So,
after weighting in the alternatives, FSecs is now defined
as boost::rational<int64_t>.

+++ src/lib/time/timecode.cpp
@@ -122,14 +122,14 @@ namespace time {
       static regex fracSecs_parser ("(?:^|[^\\./\\d\\-])(\\-?\\d+)(?:([\\-\\+]\\d+)?/(\\d+))?sec");
                                   //__no leading[./-\d]   number      [+-]  number '/' number 'sec'   
       
-      #define SUB_EXPR(N) lexical_cast<long> (match[N])
+      #define SUB_EXPR(N) lexical_cast<int> (match[N])
       smatch match;
       if (regex_search (seconds, match, fracSecs_parser))
         if (match[2].matched)
           {
             // complete spec with all parts
             FSecs fractionalPart (SUB_EXPR(2), SUB_EXPR(3));
-            long fullSeconds (SUB_EXPR(1));
+            int fullSeconds (SUB_EXPR(1));
             return grid.timeOf (fullSeconds + fractionalPart);
           }
         else
